requirements.txt
Flask==2.3.3
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.3.0
PyPDF2==3.0.1